requirements.txt
# Core Django and REST Framework
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0,<4.0.0
django-cors-headers>=4.3.0,<5.0.0

# WebSocket Support
channels>=4.0.0,<5.0.0
channels-redis>=4.1.0,<5.0.0
daphne>=4.0.0,<5.0.0

# Firebase Authentication
firebase-admin>=6.2.0,<7.0.0
pyjwt>=2.8.0,<3.0.0
cryptography>=3.4.6,<4.0.0

# File Processing
PyPDF2>=3.0.0,<4.0.0
python-docx>=0.8.11,<1.0.0

# Image Processing
Pillow>=10.0.0,<11.0.0

# Environment and Configuration
python-decouple>=3.8,<5.0.0

# Database (optional - use SQLite for development)
# psycopg2-binary>=2.9.0,<3.0.0

# Task Queue (optional - for background processing)
# celery>=5.3.0,<6.0.0
# redis>=5.0.0,<6.0.0

# File Storage (optional - for production)
# django-storages>=1.14.0,<5.0.0
# boto3>=1.34.0,<5.0.0

# Production Dependencies
gunicorn>=21.2.0,<22.0.0
dj-database-url>=2.1.0,<3.0.0
whitenoise>=6.6.0,<7.0.0
